fix `zola serve` not  respecting `preserve_dotfiles_in_output`
